refactor(nuget): Move `NuGetInspector` related extension function
Move funtions that build the bridge between NuGetInspector's and ORT's
models to a dedicated file in order for `NuGetInspector` to be
independent of ORT's model for a cleaner separation.
